+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+* [PR-625](https://github.com/itk-dev/deltag.aarhus.dk/pull/625)
+  Updated paragraph layout
 * [PR-624](https://github.com/itk-dev/deltag.aarhus.dk/pull/624)
   Fixed issue with project status
 * [PR-623](https://github.com/itk-dev/deltag.aarhus.dk/pull/623)